requirements.txt
Flask==2.3.3
yt-dlp==2023.11.16
python-dotenv==1.0.0
Flask-Limiter==3.5.0
Flask-CORS==4.0.0
gunicorn==21.2.0 